queries.py: Modified queries to filter the attributes and make several checks before adding the task. Removed integer codes to select which attributes will be used for a query. Adapted queries to use filter_params function to filter attributes whitelisted. Added conditions_query function to verify conditions over the arguments of the queries.
utils.py:
Added mode_param function, which check if the argument to get the mode is whitelisted.
Modified filter_params, added an optional parameter to select which explicit whitelist the function will use as reference to check the arguments of the request. If no second parameter is passed, the function will use the endpoint of the request, as did on the previous version.

query_handler.py:
Removed codes for filter arguments to assemble the query. Now it will filter based on the arguments passed.

Closes #35